Refactor composition inspector
